+++ src/main/java/swaglabs/reports/ExtentReporterNG.java
@@ -1,4 +1,4 @@
-package swaglabs.resources;
+package swaglabs.reports;
 
 import com.aventstack.extentreports.ExtentReports;
 import com.aventstack.extentreports.reporter.ExtentSparkReporter;
