+++ test-output/Default suite/Default test.html
@@ -57,9 +57,9 @@ function toggleAllBoxes() {
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Fri Oct 04 15:51:40 IST 2024</td>
+<td>Started on:</td><td>Tue Oct 22 11:57:22 IST 2024</td>
 </tr>
-<tr><td>Total time:</td><td>7 seconds (7341 ms)</td>
+<tr><td>Total time:</td><td>45 seconds (45562 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -75,10 +75,10 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='swaglabs.ai.tests.validateHomePageTest.validateUserLandedInSwagLabsHomePage()'><b>validateUserLandedInSwagLabsHomePage</b><br>Test class: swaglabs.ai.tests.validateHomePageTest</td>
+<td title='swaglabs.tests.validateHomePageTest.validateUserLandedInSwagLabsHomePage()'><b>validateUserLandedInSwagLabsHomePage</b><br>Test class: swaglabs.tests.validateHomePageTest</td>
 <td></td>
-<td>1</td>
-<td>swaglabs.ai.tests.validateHomePageTest@67d48005</td></tr>
+<td>32</td>
+<td>swaglabs.tests.validateHomePageTest@368247b9</td></tr>
 </table><p>
 </body>
 </html>
