+++ test-output/Default suite/Default test.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="1" hostname="LTIN438577" failures="0" tests="1" name="Default test" time="7.341" errors="0" timestamp="2024-10-04T15:51:47 IST">
-  <testcase classname="swaglabs.ai.tests.validateHomePageTest" name="validateUserLandedInSwagLabsHomePage" time="1.759"/>
-  <testcase classname="swaglabs.ai.tests.validateHomePageTest" name="validateOfAddSauceLabsBackPackIntoCartFromHomePageTest" time="0.0">
+<testsuite ignored="1" hostname="LTIN438577" failures="0" tests="1" name="Default test" time="45.562" errors="0" timestamp="2024-10-22T11:58:07 IST">
+  <testcase classname="swaglabs.tests.validateHomePageTest" name="validateUserLandedInSwagLabsHomePage" time="32.978"/>
+  <testcase classname="swaglabs.tests.validateHomePageTest" name="validateOfAddSauceLabsBackPackIntoCartFromHomePageTest" time="0.0">
     <ignored/>
   </testcase> <!-- validateOfAddSauceLabsBackPackIntoCartFromHomePageTest -->
 </testsuite> <!-- Default test -->
