+++ test-output/index.html
@@ -51,7 +51,7 @@
               <ul>
                 <li>
                   <a href="#" panel-name="test-xml-Default_suite" class="navigator-link ">
-                    <span>C:\Users\2280571\AppData\Local\Temp\testng-eclipse--1186114569\testng-customsuite.xml</span>
+                    <span>C:\Users\2280571\AppData\Local\Temp\testng-eclipse--1425687029\testng-customsuite.xml</span>
                   </a> <!-- navigator-link  -->
                 </li>
                 <li>
@@ -104,7 +104,7 @@
                     <div class="method-list-content passed suite-Default_suite">
                       <span>
                         <img src="passed.png" width="3%"/>
-                        <a href="#" panel-name="suite-Default_suite" title="swaglabs.ai.tests.validateHomePageTest" class="method navigator-link" hash-for-method="validateUserLandedInSwagLabsHomePage">validateUserLandedInSwagLabsHomePage</a> <!-- method navigator-link -->
+                        <a href="#" panel-name="suite-Default_suite" title="swaglabs.tests.validateHomePageTest" class="method navigator-link" hash-for-method="validateUserLandedInSwagLabsHomePage">validateUserLandedInSwagLabsHomePage</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                     </div> <!-- method-list-content passed suite-Default_suite -->
@@ -122,7 +122,7 @@
           <div class="suite-Default_suite-class-passed">
             <div class="main-panel-header rounded-window-top">
               <img src="passed.png"/>
-              <span class="class-name">swaglabs.ai.tests.validateHomePageTest</span>
+              <span class="class-name">swaglabs.tests.validateHomePageTest</span>
             </div> <!-- main-panel-header rounded-window-top -->
             <div class="main-panel-content rounded-window-bottom">
               <div class="method">
@@ -137,7 +137,7 @@
         </div> <!-- panel Default_suite -->
         <div panel-name="test-xml-Default_suite" class="panel">
           <div class="main-panel-header rounded-window-top">
-            <span class="header-content">C:\Users\2280571\AppData\Local\Temp\testng-eclipse--1186114569\testng-customsuite.xml</span>
+            <span class="header-content">C:\Users\2280571\AppData\Local\Temp\testng-eclipse--1425687029\testng-customsuite.xml</span>
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <pre>
@@ -146,11 +146,11 @@
 &lt;suite name=&quot;Default suite&quot; guice-stage=&quot;DEVELOPMENT&quot;&gt;
   &lt;test thread-count=&quot;5&quot; name=&quot;Default test&quot; verbose=&quot;2&quot;&gt;
     &lt;classes&gt;
-      &lt;class name=&quot;swaglabs.ai.tests.validateHomePageTest&quot;&gt;
+      &lt;class name=&quot;swaglabs.tests.validateHomePageTest&quot;&gt;
         &lt;methods&gt;
           &lt;include name=&quot;validateUserLandedInSwagLabsHomePage&quot;/&gt;
         &lt;/methods&gt;
-      &lt;/class&gt; &lt;!-- swaglabs.ai.tests.validateHomePageTest --&gt;
+      &lt;/class&gt; &lt;!-- swaglabs.tests.validateHomePageTest --&gt;
     &lt;/classes&gt;
   &lt;/test&gt; &lt;!-- Default test --&gt;
 &lt;/suite&gt; &lt;!-- Default suite --&gt;
@@ -193,13 +193,13 @@ data.addColumn('number', 'Time (ms)');
 data.addRows(1);
 data.setCell(0, 0, 0)
 data.setCell(0, 1, 'validateUserLandedInSwagLabsHomePage')
-data.setCell(0, 2, 'swaglabs.ai.tests.validateHomePageTest')
-data.setCell(0, 3, 1759);
+data.setCell(0, 2, 'swaglabs.tests.validateHomePageTest')
+data.setCell(0, 3, 32978);
 window.suiteTableData['Default_suite']= { tableData: data, tableDiv: 'times-div-Default_suite'}
 return data;
 }
               </script>
-              <span class="suite-total-time">Total running time: 1 seconds</span>
+              <span class="suite-total-time">Total running time: 32 seconds</span>
               <div id="times-div-Default_suite">
               </div> <!-- times-div-Default_suite -->
             </div> <!-- times-div -->
@@ -218,7 +218,7 @@ return data;
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <div class="ignored-class-div">
-              <span class="ignored-class-name">swaglabs.ai.tests.validateHomePageTest</span>
+              <span class="ignored-class-name">swaglabs.tests.validateHomePageTest</span>
               <div class="ignored-methods-div">
                 <span class="ignored-method-name">validateOfAddSauceLabsBackPackIntoCartFromHomePageTest</span>
                 <br/>
@@ -232,18 +232,18 @@ return data;
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <div class="chronological-class">
-              <div class="chronological-class-name">swaglabs.ai.tests.validateHomePageTest</div> <!-- chronological-class-name -->
+              <div class="chronological-class-name">swaglabs.tests.validateHomePageTest</div> <!-- chronological-class-name -->
               <div class="configuration-test before">
                 <span class="method-name">initializeDriver</span>
                 <span class="method-start">0 ms</span>
               </div> <!-- configuration-test before -->
               <div class="test-method">
                 <span class="method-name">validateUserLandedInSwagLabsHomePage</span>
-                <span class="method-start">4309 ms</span>
+                <span class="method-start">11365 ms</span>
               </div> <!-- test-method -->
               <div class="configuration-test after">
                 <span class="method-name">tearDown</span>
-                <span class="method-start">6074 ms</span>
+                <span class="method-start">44351 ms</span>
               </div> <!-- configuration-test after -->
           </div> <!-- main-panel-content rounded-window-bottom -->
         </div> <!-- panel -->
