+++ test-output/junitreports/TEST-swaglabs.tests.validateHomePageTest.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="LTIN438577" failures="0" tests="1" name="swaglabs.tests.validateHomePageTest" time="32.978" errors="0" timestamp="2024-10-22T11:58:08 IST" skipped="0">
+  <testcase classname="swaglabs.tests.validateHomePageTest" name="validateUserLandedInSwagLabsHomePage" time="32.978"/>
+  <system-out/>
+</testsuite> <!-- swaglabs.tests.validateHomePageTest -->
