+++ test-output/old/Default suite/methods-not-run.html
@@ -1,3 +1,3 @@
 <h2>Methods that were not run</h2><table>
-<tr><td>swaglabs.ai.tests.validateHomePageTest.validateOfAddSauceLabsBackPackIntoCartFromHomePageTest</td></tr>
+<tr><td>swaglabs.tests.validateHomePageTest.validateOfAddSauceLabsBackPackIntoCartFromHomePageTest</td></tr>
 </table>
