+++ test-output/old/Default suite/methods.html
@@ -1,10 +1,10 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="d6b6c7">  <td>24/10/04 15:51:40</td>   <td>0</td> <td>&nbsp;</td><td title="&gt;&gt;DriverFactory.initializeDriver()[pri:0, instance:swaglabs.ai.tests.validateHomePageTest@67d48005]">&gt;&gt;initializeDriver</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1159190947</td>   <td></td> </tr>
-<tr bgcolor="d6b6c7">  <td>24/10/04 15:51:46</td>   <td>6072</td> <td>&nbsp;</td><td title="&lt;&lt;DriverFactory.tearDown()[pri:0, instance:swaglabs.ai.tests.validateHomePageTest@67d48005]">&lt;&lt;tearDown</td> 
-<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1159190947</td>   <td></td> </tr>
-<tr bgcolor="9fe9b6">  <td>24/10/04 15:51:44</td>   <td>4306</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="validateHomePageTest.validateUserLandedInSwagLabsHomePage()[pri:0, instance:swaglabs.ai.tests.validateHomePageTest@67d48005]">validateUserLandedInSwagLabsHomePage</td> 
-  <td>main@1159190947</td>   <td></td> </tr>
+<tr bgcolor="daa2b3">  <td>24/10/22 11:57:22</td>   <td>0</td> <td>&nbsp;</td><td title="&gt;&gt;DriverFactory.initializeDriver()[pri:0, instance:swaglabs.tests.validateHomePageTest@368247b9]">&gt;&gt;initializeDriver</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1390988828</td>   <td></td> </tr>
+<tr bgcolor="daa2b3">  <td>24/10/22 11:58:06</td>   <td>44346</td> <td>&nbsp;</td><td title="&lt;&lt;DriverFactory.tearDown()[pri:0, instance:swaglabs.tests.validateHomePageTest@368247b9]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@1390988828</td>   <td></td> </tr>
+<tr bgcolor="c8eb65">  <td>24/10/22 11:57:33</td>   <td>11360</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="validateHomePageTest.validateUserLandedInSwagLabsHomePage()[pri:0, instance:swaglabs.tests.validateHomePageTest@368247b9]">validateUserLandedInSwagLabsHomePage</td> 
+  <td>main@1390988828</td>   <td></td> </tr>
 </table>
