+++ test-output/testng-results.xml
@@ -2,24 +2,24 @@
 <testng-results ignored="1" total="2" passed="1" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2024-10-04T15:51:40 IST" name="Default suite" finished-at="2024-10-04T15:51:47 IST" duration-ms="7341">
+  <suite started-at="2024-10-22T11:57:22 IST" name="Default suite" finished-at="2024-10-22T11:58:07 IST" duration-ms="45562">
     <groups>
     </groups>
-    <test started-at="2024-10-04T15:51:40 IST" name="Default test" finished-at="2024-10-04T15:51:47 IST" duration-ms="7341">
-      <class name="swaglabs.ai.tests.validateHomePageTest">
-        <test-method is-config="true" signature="initializeDriver()[pri:0, instance:swaglabs.ai.tests.validateHomePageTest@67d48005]" started-at="2024-10-04T15:51:40 IST" name="initializeDriver" finished-at="2024-10-04T15:51:44 IST" duration-ms="4270" status="PASS">
+    <test started-at="2024-10-22T11:57:22 IST" name="Default test" finished-at="2024-10-22T11:58:07 IST" duration-ms="45562">
+      <class name="swaglabs.tests.validateHomePageTest">
+        <test-method is-config="true" signature="initializeDriver()[pri:0, instance:swaglabs.tests.validateHomePageTest@368247b9]" started-at="2024-10-22T11:57:22 IST" name="initializeDriver" finished-at="2024-10-22T11:57:33 IST" duration-ms="11334" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- initializeDriver -->
-        <test-method signature="validateUserLandedInSwagLabsHomePage()[pri:0, instance:swaglabs.ai.tests.validateHomePageTest@67d48005]" started-at="2024-10-04T15:51:44 IST" name="validateUserLandedInSwagLabsHomePage" finished-at="2024-10-04T15:51:46 IST" duration-ms="1759" status="PASS">
+        <test-method signature="validateUserLandedInSwagLabsHomePage()[pri:0, instance:swaglabs.tests.validateHomePageTest@368247b9]" started-at="2024-10-22T11:57:33 IST" name="validateUserLandedInSwagLabsHomePage" finished-at="2024-10-22T11:58:06 IST" duration-ms="32978" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- validateUserLandedInSwagLabsHomePage -->
-        <test-method is-config="true" signature="tearDown()[pri:0, instance:swaglabs.ai.tests.validateHomePageTest@67d48005]" started-at="2024-10-04T15:51:46 IST" name="tearDown" finished-at="2024-10-04T15:51:47 IST" duration-ms="1262" status="PASS">
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:swaglabs.tests.validateHomePageTest@368247b9]" started-at="2024-10-22T11:58:06 IST" name="tearDown" finished-at="2024-10-22T11:58:07 IST" duration-ms="1205" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- tearDown -->
-      </class> <!-- swaglabs.ai.tests.validateHomePageTest -->
+      </class> <!-- swaglabs.tests.validateHomePageTest -->
     </test> <!-- Default test -->
   </suite> <!-- Default suite -->
 </testng-results>
